refactor: 💡 move prometheus code to libcommon
it includes StepProfiler to profile the code
